Extract runtime/tracing/throttle bring-up from lib.rs into runtime_setup
The crate root carried ~1100 lines of process-startup plumbing (tokio runtime
build, tracing-subscriber install, throttle replenishers, adaptive
metadata-ops control-loop wiring, runtime-stats collection, trace/histogram
file helpers) mixed with the shared types, the progress driver, the op
wrappers, and the run() entry point. Move that cluster into a new runtime_setup
module, leaving a 714-line root (was 1805).

Pure code motion, no behavior change: function bodies are verbatim; the only
edits are the minimal pub(crate) widenings the compiler requires for the
cross-module calls (PROGRESS/PBAR/REMOTE_RUNTIME_STATS, two logger-state
helpers, and the moved fns run() calls plus TracingGuards), plus pub use
re-exports so common::collect_runtime_stats / generate_debug_log_filename /
generate_trace_filename are unchanged.
